Revert "Fix Theme Issues"
