+++ .github/workflows/ci.yml
@@ -6,11 +6,6 @@ on:
   pull_request:
     branches: [ main ]
 
-permissions:
-  contents: read
-  pull-requests: write
-  issues: write
-
 env:
   REGISTRY: ghcr.io
   IMAGE_NAME: ${{ github.repository }}
@@ -164,12 +159,11 @@ jobs:
     name: Deploy Preview (PR only)
     runs-on: ubuntu-latest
     needs: build-and-test-docker
-    if: github.event_name == 'pull_request' && github.event.pull_request.head.repo.full_name == github.repository
+    if: github.event_name == 'pull_request'
     
     steps:
       - name: Comment PR with preview info
         uses: actions/github-script@v7
-        continue-on-error: true
         with:
           script: |
             github.rest.issues.createComment({
