+++ lib/rateLimiter.ts
@@ -1,4 +1,4 @@
-import { LRUCache } from "lru-cache";
+import LRUCache from "lru-cache";
 
 export interface RateLimiterConfig {
   maxRequests: number;
